requirements.txt
numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.15"
opencv-python-headless==4.11.0.86 ; python_version >= "3.11" and python_version < "3.15"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "3.15"
ultralytics==8.3.233 ; python_version >= "3.11" and python_version < "3.15"

